Move `NodeId` type into `id` module
